Allow ntrip_daemon to terminate correctly
