+++ package/ntrip_daemon/src/ntrip_settings.h
@@ -17,5 +17,6 @@
 
 void ntrip_init(settings_ctx_t *settings_ctx);
 bool ntrip_reconnect(void);
+void ntrip_stop_processes(void);
 
 #endif
