+++ package/skylark_daemon/src/skylark_settings.c
@@ -12,7 +12,9 @@
 
 #include <unistd.h>
 #include <errno.h>
+
 #include <sys/stat.h>
+
 #include <libpiksi/logging.h>
 #include <libpiksi/util.h>
